refactor: move named impls out of generated
